Redirecting invocations with hardcoded paths
red can now be configured to redirect the invocation of certain tools.
By specifying an alternative path for `tool`, red will redirect any
invocation of `/usr/bin/tool` or `/usr/sbin/tool` or `/bin/tool` to that
path.

Specifying which tools to redirect, and alternate paths, is done at
compile time for efficiency. Users can pass the flag
-DBEAR_TOOL=/new/path to redirect absolute-path invocations of TOOL to
/new/path. The list of supported tools is in libred/CMakeLists.txt.

This commit does NOT deal with the situation where a build script
invokes a tool without a relative path, but clobbers the PATH variable
so that an invocation of `ar` resolves to `/usr/bin/ar` anyway.

The PID of the execing process and the name of the tool whose absolute
path was invoked is written out to a temporary file.