Inject a directory into PATH if it isn't there
Upon each invocation, red can now scan the environment for a $PATH
variable and correct it if it does not point to the right directories.

In order to configure this behaviour, red must be built with the
-DRED_FIRST_PATH flag set to the first directory that is expected to be
in $PATH.

If that directory is not the first directory in $PATH, then red shall
modify $PATH so that it is, and additionally dump a warning message. The
same thing happens if $PATH is empty.

If the environment does not contain $PATH at all, then red dumps a
diagnostic message (but does not change the environment).